Change OUTPUT_DECIMALS to uint8_t.

+++ include/qx/CompileTimeConfiguration.hpp
@@ -10,7 +10,7 @@ namespace qx::config {
 static constexpr double EPS = 0.000000000001;
 
 // Number of decimals in output
-static constexpr std::uint64_t const OUTPUT_DECIMALS = 8;
+static constexpr std::uint8_t const OUTPUT_DECIMALS = 8;
 
 // How many gates between cleaning the zeros in the sparse array
 static constexpr std::uint64_t ZERO_CYCLE_SIZE = 100;
